重构歌词(https://github.com/Henry-ZHR/APlayer.git)

+++ app/src/main/java/remix/myplayer/bean/netease/NSongSearchResponse.kt
@@ -5,7 +5,7 @@ data class NSongSearchResponse(val code: Int = 0,
 
   data class SearchActualData(val songs: List<SearchInnerData?>? = null) {
 
-    data class SearchInnerData(val id: Int,
+    data class SearchInnerData(val id: Long,
                                val album: SearchAlbumData? = null,
                                val score: Int) {
